+++ wyndpy/parse/__init__.py
@@ -1,0 +1,5 @@
+"""Parse-role protocol."""
+
+from wyndpy.parse.protocol import ParserProtocol
+
+__all__ = ["ParserProtocol"]