+++ wyndpy/search/__init__.py
@@ -1,0 +1,5 @@
+"""Search-role protocol."""
+
+from wyndpy.search.protocol import SearcherProtocol
+
+__all__ = ["SearcherProtocol"]